### 2023-10-03 #5922 @editorialbot
[PRE REVIEW]: Redflag: machine learning safety by design
**Submitting author:** @kwinkunks (<a href="http://orcid.org/0000-0002-4054-8295">Matt Hall</a>)
**Repository:** https://github.com/scienxlab/redflag
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.4.1
