### 2023-10-03 5eb9b45
[issue-83] - Fixed utils.docstring_from() decorator
