+++ .github/workflows/publish-docs.yml
@@ -27,7 +27,7 @@ jobs:
         make html
 
     - name: Publish docs
-      uses: JamesIves/github-pages-deploy-action@v4.4.3
+      uses: JamesIves/github-pages-deploy-action@v4
       with:
         branch: gh-pages
         folder: docs/_build/html
