### 2023-12-03 scienxlab/redflag#89 @jsheunis
JOSS paper feedback
Hi @kwinkunks

The JOSS paper itself is nicely written, well done. Here are some minor feedback points:

- Based on the checklist "Do the authors describe how this software compares to other commonly-used packages", you do provide examples of similar packages in lines 61-65, but you don't elaborate on how `redflag` compares to these. I would add a couple of sentences to explain differences.
- Typo in line 50:
   > There are two kinds of scikit-learn transformer

   --> There are two kinds of scikit-learn transformers:
- Line 64-65, typo, and more info needed:
   > Finally, [evidently](https://github.com/evidentlyai/evidently) provides on a Jupyter interface with lots of plots.

   --> Finally, [evidently](https://github.com/evidentlyai/evidently) provides ~on~ a Jupyter interface with lots of plots.
   Also, "lots of plots" sounds very vague, perhaps you could provide a bit more detail about what this package provides.

---
Ping https://github.com/openjournals/joss-reviews/issues/6065
