### 2023-12-03 scienxlab/redflag#90 @jsheunis
Problem installing `[dev]` option
I'm installing via `pip` in a virtual environment on macOS:

```
> pip install redflag[dev]
zsh: no matches found: redflag[dev]
```

On the other hand `pip install redflag` works fine.

---
Ping https://github.com/openjournals/joss-reviews/issues/6065
