### 2023-12-03 scienxlab/redflag#91 @jsheunis
Example csv file inaccessible
Running through the example usage in the docs (https://scienxlab.org/redflag/_notebooks/Basic_usage.html), the file at https://geocomp.s3.amazonaws.com/data/Panoma_training_data.csv seems to be inaccessible.

```
This XML file does not appear to have any style information associated with it. The document tree is shown below.
<Error>
<Code>AllAccessDisabled</Code>
<Message>All access to this object has been disabled</Message>
<RequestId>0W430R97NCX2H4AT</RequestId>
<HostId>tzZmU3uCbN0V3jNgZtk8q83TvNVt7DY5d6n0JlIRwDpw1rRxytuZ5tnVX84Owjg0AZYHNVQMk/I=</HostId>
</Error>
```

It would be good to use a publicly available file so that users can run the example on their own.

---
Ping https://github.com/openjournals/joss-reviews/issues/6065
