### 2023-12-03 scienxlab/redflag#92 @jsheunis
CI testing environment
I see the CI only runs on ubuntu: https://github.com/scienxlab/redflag/blob/main/.github/workflows/build-test.yml

I wouldn't say its absolutely necessary, but it would probably be a good idea to run tests on all major operating systems. If not, I suggest stating that in the documentation.

---
Ping https://github.com/openjournals/joss-reviews/issues/6065
