### 2023-12-03 scienxlab/redflag#93 @jsheunis
Lengthy one-page API docs
The Python API documentation is quite lengthy for a single page (https://scienxlab.org/redflag/redflag.html#), and requires lots of scrolling / searching to find specific content. I suggest splitting it up into one page per module.

---
Ping https://github.com/openjournals/joss-reviews/issues/6065
