### 2023-12-03 scienxlab/redflag#94 @jsheunis
Minimal context in docs
The docs contain a single paragraph explaining the context within which this package operates: https://scienxlab.org/redflag/index.html

I think it would make sense to provide a bit more information to position the software within the context of the ML field so that users can develop a more intuitive understanding of its benefits and use cases. Relatedly, the JOSS checklist states:

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

This is covered nicely in the JOSS paper itself, and perhaps some of this content can be worked into the docs. I actually like the current single paragraph intro in the docs index page before jumping into the quickstart, so perhaps a more detailed overview can be provided on a separate page and referenced from the index page.

Overall, the docs are great. I appreciate the attention given to this. Also, +1 on Furo :)

---
Ping https://github.com/openjournals/joss-reviews/issues/6065
