### 2023-12-09 3895094
update optional install info
Fixes #90
