### 2023-12-09 scienxlab/redflag#90 @kwinkunks
Good catch! I think `zsh` interprets the option as regex, so you have to do `pip install "reflag[dev]"`, with the quotes. And I think that works with most other shells too.

Updated docs with [3895094](https://github.com/scienxlab/redflag/commit/38950949409312977b0acabe0da67702e81ebeaf)
