### 2023-12-09 scienxlab/redflag#91 @kwinkunks
Ack, good one, thanks — I shut down an AWS account and forgot about the S3.

Moved the files to a new repo, added a data license (which the file didn't have before) and changed the links in the docs to the new file.
