### 2023-12-09 c52d3ce
Update CI and test on all OS's
Fixes #92
