### 2023-12-09 scienxlab/redflag#92 @kwinkunks
Great suggestion, thank you.

Done in [c52d3ce](https://github.com/scienxlab/redflag/commit/c52d3ceafb3f57c05d1c6fa51e7b28390d531f6a)
