### 2023-12-09 046d758
Separate API docs into modules
Fixes #93
