+++ docs/redflag.rst
@@ -4,85 +4,19 @@ redflag package
 Submodules
 ----------
 
-redflag.distributions module
-----------------------------
+.. toctree::
+   :maxdepth: 4
 
-.. automodule:: redflag.distributions
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-redflag.imbalance module
-------------------------
-
-.. automodule:: redflag.imbalance
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-redflag.importance module
--------------------------
-
-.. automodule:: redflag.importance
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-redflag.independence module
----------------------------
-
-.. automodule:: redflag.independence
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-redflag.markov module
----------------------
-
-.. automodule:: redflag.markov
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-redflag.outliers module
------------------------
-
-.. automodule:: redflag.outliers
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-redflag.pandas module
----------------------
-
-.. automodule:: redflag.pandas
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-redflag.sklearn module
-----------------------
-
-.. automodule:: redflag.sklearn
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-redflag.target module
----------------------
-
-.. automodule:: redflag.target
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-redflag.utils module
---------------------
-
-.. automodule:: redflag.utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
+   redflag.distributions
+   redflag.imbalance
+   redflag.importance
+   redflag.independence
+   redflag.markov
+   redflag.outliers
+   redflag.pandas
+   redflag.sklearn
+   redflag.target
+   redflag.utils
 
 Module contents
 ---------------
