### 2023-12-09 scienxlab/redflag#93 @kwinkunks
Great idea, I need to learn more about the `apidoc` plugin.

Fixed in 046d75812abc87b794e320067aa44a07006bb941
