### 2023-12-09 5e534d8
Added 'what is' section to docs
Also tweaked some of the other pages to use more consistent styling.
Fixes #94.
