### 2023-12-09 scienxlab/redflag#94 @kwinkunks
Thanks for the kind words, appreciated.

Totally agree that more scene-setting was needed -- I often look for this myself in other packages. I've tried to address this with a new 'What is redflag?' page in the docs. I have self-plagiarized it from the draft paper -- I hope that's okay? Happy to re-word if it seems like too much of a shortcut.

Proposed fix in a525174aa6937c351b871a9053904fbfc279c57f
