### 2023-12-09 scienxlab/redflag#89 @kwinkunks
Thanks for the feedback. I think I addressed these in the latest draft.

- Added this sentence:

  > In comparison, _Redflag_ is easier to set up and use
than `great_expectations` and `pandera`, and while it is compatible with
Pandas DataFrames and Jupyter it does not depend on them.

- Reworded sentence:

  > The `scikit-learn` transformers are of two kinds:

- Fixed typo and clarified:

  > Finally,
[`evidently`](https://github.com/evidentlyai/evidently) provides a graphical
dashboard for Jupyter.
