### 2023-12-09 0d24093
Ignore dtype in doctests on Windows
Fixes #96
