+++ tests/README.md
@@ -9,3 +9,9 @@ The Jupyter Notebooks in `docs/notebooks` are currently not run as part of the t
 Test options are in `pyproject.toml`, so to run the tests: clone the repo, install the dev dependencies (e.g. with `"pip install .[dev]`") and do this from the root directory:
 
     pytest
+
+
+## A note about NumPy dtypes
+
+Owing to an idiosyncracy of 64-bit Windows machines, which count a 'long' int as 32-bit not 64, I have stopped `doctest` from comparing any `dtype=int64` or similar in test outputs. This is done by the custom `doctest.OutputChecker` in `tests/conftest.py`. It only runs on Windows
+machines (e.g. in the CI matrix).
