+++ _sources/_notebooks/Using_redflag_with_Pandas.ipynb.txt
@@ -1,1 +1,1 @@
-{"cells": [{"cell_type": "markdown", "id": "a8d12712-5c7b-4acb-bb8b-e73efcb9b5dc", "metadata": {}, "source": ["# \ud83d\udea9 Using `redflag` with Pandas\n", "\n", "As well as using `redflag`'s functions directly (see [`Basic_usage.ipynb`](`./Basic_usage.ipynb`)), or with `sklearn` (see `Using_redflag_with_Pandas.ipynb`), `redflag` has some Pandas 'accessors' that give you access to some `redflag` functions almost as if they were methods on Pandas objects.\n", "\n", "The best way to get the idea is to look at an example.\n", "\n", "First, even though we may not use it directly, we have to import `redflag` to get access to its functions. As long as you have `pandas` installed, it will register the accessors."]}, {"cell_type": "code", "execution_count": 1, "id": "77aa7f67-0bc7-48e9-87f4-183aa2dc2c35", "metadata": {}, "outputs": [{"data": {"text/plain": ["'0.4.1.dev4+ga088506.d20231002'"]}, "execution_count": 1, "metadata": {}, "output_type": "execute_result"}], "source": ["import redflag as rf\n", "\n", "rf.__version__"]}, {"cell_type": "code", "execution_count": 2, "id": "3dbcf6e1-1cb5-4ca5-b64a-bc1d9e7b174f", "metadata": {}, "outputs": [{"data": {"text/html": ["<div>\n", "<style scoped>\n", "    .dataframe tbody tr th:only-of-type {\n", "        vertical-align: middle;\n", "    }\n", "\n", "    .dataframe tbody tr th {\n", "        vertical-align: top;\n", "    }\n", "\n", "    .dataframe thead th {\n", "        text-align: right;\n", "    }\n", "</style>\n", "<table border=\"1\" class=\"dataframe\">\n", "  <thead>\n", "    <tr style=\"text-align: right;\">\n", "      <th></th>\n", "      <th>Well Name</th>\n", "      <th>Depth</th>\n", "      <th>Formation</th>\n", "      <th>RelPos</th>\n", "      <th>Marine</th>\n", "      <th>GR</th>\n", "      <th>ILD</th>\n", "      <th>DeltaPHI</th>\n", "      <th>PHIND</th>\n", "      <th>PE</th>\n", "      <th>Facies</th>\n", "      <th>LATITUDE</th>\n", "      <th>LONGITUDE</th>\n", "      <th>ILD_log10</th>\n", "      <th>Lithology</th>\n", "      <th>RHOB</th>\n", "      <th>Mineralogy</th>\n", "      <th>Siliciclastic</th>\n", "    </tr>\n", "  </thead>\n", "  <tbody>\n", "    <tr>\n", "      <th>0</th>\n", "      <td>SHRIMPLIN</td>\n", "      <td>851.3064</td>\n", "      <td>A1 SH</td>\n", "      <td>1.000</td>\n", "      <td>1</td>\n", "      <td>77.45</td>\n", "      <td>4.613176</td>\n", "      <td>9.9</td>\n", "      <td>11.915</td>\n", "      <td>4.6</td>\n", "      <td>3.0</td>\n", "      <td>37.978076</td>\n", "      <td>-100.987305</td>\n", "      <td>0.664</td>\n", "      <td>siltstone</td>\n", "      <td>2393.499945</td>\n", "      <td>siliciclastic</td>\n", "      <td>True</td>\n", "    </tr>\n", "    <tr>\n", "      <th>1</th>\n", "      <td>SHRIMPLIN</td>\n", "      <td>851.4588</td>\n", "      <td>A1 SH</td>\n", "      <td>0.979</td>\n", "      <td>1<
[…]
