+++ _sources/_notebooks/Tutorial.ipynb.txt
@@ -1,1 +1,1 @@
-{"cells": [{"cell_type": "markdown", "metadata": {}, "source": ["# \ud83d\udea9 Tutorial\n", "\n", "We're going to look at some features of `redflag`, a library for helping find problems in machine learning pipelines.\n", "\n", "You'll need the following packages to run the code in this tutorial:\n", "\n", "- `redflag`\n", "- `pandas`\n", "- `seaborn`\n", "\n", "\n", "## A simple ML workflow\n", "\n", "First, let's see how we can burn ourselves:"]}, {"cell_type": "code", "execution_count": 1, "metadata": {}, "outputs": [{"data": {"text/plain": ["array(['ss', 'ss', 'ss', 'ms', 'ms', 'ss'], dtype='<U2')"]}, "execution_count": 1, "metadata": {}, "output_type": "execute_result"}], "source": ["X = [[19], [23], [35], [64], [59], [31]]  # The smallest gamma-ray log.\n", "y = ['ss', 'ss', 'ss', 'ms', 'ms', 'ss']\n", "\n", "from sklearn.svm import SVC\n", "\n", "clf = SVC(kernel='linear')\n", "clf.fit(X, y)\n", "clf.predict(X)"]}, {"cell_type": "markdown", "metadata": {}, "source": ["So far so good. We're predicting on the training data, but everything is at least working.\n", "\n", "Now someone tells us we should scale our training data."]}, {"cell_type": "code", "execution_count": 2, "metadata": {}, "outputs": [{"data": {"text/plain": ["array(['ms', 'ms', 'ms', 'ms', 'ms', 'ms'], dtype='<U2')"]}, "execution_count": 2, "metadata": {}, "output_type": "execute_result"}], "source": ["from sklearn.preprocessing import StandardScaler\n", "\n", "scaler = StandardScaler()\n", "scaler.fit(X)\n", "X_scaled = scaler.transform(X)\n", "\n", "clf.fit(X_scaled, y)\n", "clf.predict(X)  # <-- Oops, we predicted on unscaled data."]}, {"cell_type": "markdown", "metadata": {}, "source": ["Easily done. There are lots of people on Stack Overflow and Cross Validated wondering why all their predictions are the same. It's often because they've done something like this.\n", "\n", "Even easier is this common pattern:"]}, {"cell_type": "code", "execution_count": 3, "metadata": {}, "outputs": [{"data": {"text/plain": ["array(['ss', 'ss'], dtype='<U2')"]}, "execution_count": 3, "metadata": {}, "output_type": "execute_result"}], "source": ["from sklearn.model_selection import train_test_split\n", "\n", "scaler = StandardScaler()\n", "scaler.fit(X)\n", "\n", "X_train, X_test, y_train, y_test = train_test_split(X, y)\n", "\n", "X_train_scaled 
[…]
