### 2023-12-10 50e264a
Shorten badges
