+++ .github/workflows/publish-docs.yml
@@ -1,4 +1,4 @@
-name: Publish docs
+name: Docs
 
 on:
   workflow_call:
