### 2023-12-17 6fbbad2
Use random seed in dummy tests
Fixes #95
