### 2023-12-17 1a995b4
Improve is_correlated docstring
Part of #97
