### 2023-12-17 896e8c2
Improve feature_importances function
Part of #97. The goal is to make the function more explainable.
Moved the aggregation function to utils.py where it can be
inspected more easily. By default it's a simple sum, but
for the feature importance measure it normalizes before
summing over features, then normalizes the result so the
importances sum to 1. Borda counts are an option too but
I think scores are best for importance, to preserve the
magnitudes of the coefficients.

Also switched to LinearRegression instead of Lasso, which
should really be fitted for alpha, see #98.
