### 2023-12-21 dcb61a5
Intermediate towards #97
