### 2024-04-09 #6065 @kwinkunks
Hello @mbobra -- thank you for checking in. Sorry to be so slow. I must admit, I got a bit crushed by item 4 in the 'final' issue https://github.com/scienxlab/redflag/issues/97 -- probably I blew it out of proportion. If it's okay with you, I will review where I am and close that issue -- then I'll come back here ASAP.
