+++ src/redflag/__init__.py
@@ -13,17 +13,15 @@ from .independence import *
 from .importance import *
 from .outliers import *
 
+from .mypackage import *
+from .cli import cli
 
-from pkg_resources import get_distribution, DistributionNotFound
-try:
-    VERSION = get_distribution(__name__).version
-except DistributionNotFound:
-    try:
-        from ._version import version as VERSION
-    except ImportError:
-        raise ImportError(
-            "Failed to find (autogenerated) _version.py. "
-            "This might be because you are installing from GitHub's tarballs, "
-            "use the PyPI ones."
-            )
-__version__ = VERSION
+# It used to be conventional to define a __version__ attribute.
+# However, it is now considered best practice to get version
+# information from the package metadata directly, eg by using
+# importlib.metadata.version (see below).
+#
+# This will be deprecated in v0.5.0 but for now we do this:
+#
+from importlib.metadata import version
+__version__ = version(__package__ or __name__)
