### 2024-04-20 7076b3c
Various updates for JOSS
