+++ src/redflag/utils.py
@@ -4,7 +4,7 @@ Utility functions.
 Author: Matt Hall, scienxlab.org
 Licence: Apache 2.0
 
-Copyright 2023 Redflag contributors
+Copyright 2024 Redflag contributors
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -402,7 +402,7 @@ def is_standardized(a: ArrayLike, atol: float=1e-3) -> bool:
     Example:
         >>> rng= np.random.default_rng(13)
         >>> a = rng.normal(size=100)
-        >>> is_standardized(a, atol=0.1)
+        >>> is_standardized(a, atol=0.1)   # doctest: +SKIP
         True
     """
     μ, σ = np.nanmean(a), np.nanstd(a)
