+++ docs/post_process_html.py
@@ -11,9 +11,9 @@ def simplify_credits(html):
     pattern = re.compile(s)
     html = pattern.sub(r'', html)
 
-    s = r'Copyright &#169; 2023, The Redflag Authors'
+    s = r'Copyright &#169; 2024, The Redflag Authors'
     pattern = re.compile(s)
-    new_s = '&#169; 2023, The Redflag Authors | <a href="https://creativecommons.org/licenses/by/4.0/">CC BY</a>'
+    new_s = '&#169; 2024, The Redflag Authors | <a href="https://creativecommons.org/licenses/by/4.0/">CC BY</a>'
     html = pattern.sub(new_s, html)
 
     return html
