### 2024-04-21 92d9c6f
Switch default outlier method, fixes #101
