+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 
 - This is a minor release making more changes to the tests and documentation in reponse to the review process for [the submission](https://joss.theoj.org/papers/e1ca575ec0c5344144f87176539ef547) to JOSS (see below).
 - In particular, see the following issue: [#97](https://github.com/scienxlab/redflag/issues/97)
-- Changed the method of handling dynamic versionin. For now the package `__version__` attribute is still defined, but it is deprecated and will be removed in `0.5.0`. Use `from importlib.metadata.version('redflag')` to get the version information instead.
+- Changed the method of handling dynamic versioning. For now the package `__version__` attribute is still defined, but it is deprecated and will be removed in `0.5.0`. Use `from importlib.metadata.version('redflag')` to get the version information instead.
+- Changed the default `get_outliers()` method from isolation forest (`'iso'`) to Mahalanobis (`'mah'`) to match other functions, eg `has_outliers()` and the `sklearn` pipeline object.
 
 
 ## 0.4.2, 10 December 2023
